refactor: Make IdentifiableItemInfo be an item property
While I started working on gear encoding again, I realized, I am going to need to access `IdentifiableItemInfo` not just for tooltips, but for encoding any gear-like item. I originally just wanted to move the class from the tooltip handler, but while searching for the correct model, I realized, we could just make it an item property.
