+++ common/src/main/java/com/wynntils/models/items/items/game/TomeItem.java
@@ -4,13 +4,20 @@
  */
 package com.wynntils.models.items.items.game;
 
+import com.wynntils.models.character.type.ClassType;
 import com.wynntils.models.gear.type.GearTier;
 import com.wynntils.models.items.properties.GearTierItemProperty;
+import com.wynntils.models.items.properties.IdentifiableItemProperty;
 import com.wynntils.models.rewards.type.TomeInfo;
 import com.wynntils.models.rewards.type.TomeInstance;
+import com.wynntils.models.stats.type.StatActualValue;
+import com.wynntils.models.stats.type.StatPossibleValues;
+import com.wynntils.models.stats.type.StatType;
+import com.wynntils.utils.type.Pair;
+import java.util.List;
 import java.util.Optional;
 
-public class TomeItem extends GameItem implements GearTierItemProperty {
+public class TomeItem extends GameItem implements GearTierItemProperty, IdentifiableItemProperty {
     private final TomeInfo tomeInfo;
     private final TomeInstance tomeInstance;
     private final int rerolls;
@@ -36,6 +43,52 @@ public class TomeItem extends GameItem implements GearTierItemProperty {
     @Override
     public GearTier getGearTier() {
         return tomeInfo.tier();
+    }
+
+    @Override
+    public String getName() {
+        return tomeInfo.name();
+    }
+
+    @Override
+    public ClassType getRequiredClass() {
+        // Tomes are not class-specific
+        return null;
+    }
+
+    @Override
+    public List<StatType> getVariableStats() {
+        return tomeInfo.variableStats().stream().map(Pair::a).toList();
+    }
+
+    @Override
+    public List<StatActualValue> getIdentifications() {
+        return tomeInstance.identifications();
+    }
+
+    @Override
+    public List<StatPossibleValues> getPossibleValues() {
+        return tomeInfo.variableStats().stream().map(Pair::b).toList();
+    }
+
+    @Override
+    public boolean hasOverallValue() {
+        return tomeInstance != null && tomeInstance.hasOverallValue();
+    }
+
+    @Override
+    public boolean isPerfect() {
+        return tomeInstance != null && tomeInstance.isPerfect();
+    }
+
+    @Override
+    public boolean isDefective() {
+        return tomeInstance != null && tomeInstance.isDefective();
+    }
+
+    @Override
+    public float getOverallPercentage() {
+        return tomeInstance != null ? tomeInstance.getOverallPercentage() : 0.0f;
     }
 
     @Override
