+++ common/src/main/java/com/wynntils/models/items/properties/IdentifiableItemProperty.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.handlers.tooltip.type;
+package com.wynntils.models.items.properties;
 
 import com.wynntils.models.character.type.ClassType;
 import com.wynntils.models.stats.type.StatActualValue;
@@ -13,7 +13,7 @@ import java.util.List;
 /**
  * Represents an item that can be identified, and provides the information needed to build a tooltip for it.
  */
-public interface IdentifiableItemInfo {
+public interface IdentifiableItemProperty {
     String getName();
 
     ClassType getRequiredClass();
